REFACTORED- Design Pattern DAO And SOLID Principles Followed

+++ src/main/java/com/bridgelabz/censusanalyser/model/IndiaCensusCSV.java
@@ -10,7 +10,7 @@ public class IndiaCensusCSV {
     public int population;
 
     @CsvBindByName(column = "AreaInSqKm", required = true)
-    public int areaInSqKm;
+    public double areaInSqKm;
 
     @CsvBindByName(column = "DensityPerSqKm", required = true)
     public int densityPerSqKm;
